requirements.txt
databricks-connect==13.3.* # Must match Runtime Maj/Min